dodati servisi, odradjena metoda findProfessorCourses

+++ eObrazovanje/src/main/java/ftn/tseo/eEducation/repository/CourseRepository.java
@@ -10,7 +10,9 @@ import ftn.tseo.eEducation.model.Course;
 
 
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
+	
+	
 
 	
 }
